Creating Python script for getting data for db (#1)
== About ==
This pull request adds a Python script that interacts with the https://jikan.moe/ Web API to retrieve data. The script then organizes this data into separate .csv files, each corresponding to different tables needed for the project. These tables include:

1. Manga
2. Author
3. Demographic
4. Reviews
5. Genres
6. Theme

It's important to note that the primary key for most of these tables is the bookid, which is provided by the https://jikan.moe/ API.

== Notes ==
* Program may take some time to create a db, need to sleep for 4 seconds to not hit rate limit
* Already created 100 row table with everything, can just load that into db